+++ packages/apify/src/configuration.ts
@@ -106,6 +106,7 @@ export interface ConfigurationOptions extends CoreConfigurationOptions {
  * `proxyPort` | `APIFY_PROXY_PORT` | `8000`
  * `proxyStatusUrl` | `APIFY_PROXY_STATUS_URL` | `'http://proxy.apify.com'`
  * `userId` | `APIFY_USER_ID` | -
+ * `userIsPaying` | `APIFY_USER_IS_PAYING` | -
  * `xvfb` | `APIFY_XVFB` | -
  * `standbyPort` | `ACTOR_STANDBY_PORT` | `4321`
  * `standbyUrl` | `ACTOR_STANDBY_URL` | -
@@ -114,7 +115,7 @@ export interface ConfigurationOptions extends CoreConfigurationOptions {
  */
 export class Configuration extends CoreConfiguration {
     /** @inheritDoc */
-    // eslint-disable-next-line no-use-before-define -- Self-reference
+
     static override globalConfig?: Configuration;
 
     // maps environment variables to config keys (e.g. `APIFY_MEMORY_MBYTES` to `memoryMbytes`)
@@ -151,6 +152,7 @@ export class Configuration extends CoreConfiguration {
         APIFY_CONTAINER_PORT: 'containerPort',
         APIFY_CONTAINER_URL: 'containerUrl',
         APIFY_USER_ID: 'userId',
+        APIFY_USER_IS_PAYING: 'userIsPaying',
         APIFY_PROXY_HOSTNAME: 'proxyHostname',
         APIFY_PROXY_PASSWORD: 'proxyPassword',
         APIFY_PROXY_STATUS_URL: 'proxyStatusUrl',