+++ packages/apify/src/key_value_store.ts
@@ -1,9 +1,10 @@
 import type { StorageManagerOptions } from '@crawlee/core';
 import { KeyValueStore as CoreKeyValueStore } from '@crawlee/core';
+import { KeyValueStoreClient as RemoteKeyValueStoreClient } from 'apify-client';
 
 import { createHmacSignature } from '@apify/utilities';
 
-import type { Configuration } from './configuration';
+import type { Configuration } from './configuration.ts';
 
 // @ts-ignore newer crawlee versions already declare this method in core
 const { getPublicUrl } = CoreKeyValueStore.prototype;
@@ -18,7 +19,13 @@ export class KeyValueStore extends CoreKeyValueStore {
      */
     override getPublicUrl(key: string): string {
         const config = this.config as Configuration;
-        if (!config.get('isAtHome') && getPublicUrl) {
+
+        const isLocalStore = !(
+            // eslint-disable-next-line dot-notation
+            (this['client'] instanceof RemoteKeyValueStoreClient)
+        );
+
+        if (isLocalStore && getPublicUrl) {
             return getPublicUrl.call(this, key);
         }